+++ dyndis/candidate.py
@@ -1,8 +1,21 @@
 from inspect import signature, Parameter
 from itertools import chain, product, permutations
-from typing import Union, Callable, get_type_hints, Any, Tuple, Collection, TypeVar
+from typing import Union, Callable, get_type_hints, Any, Tuple, Collection, TypeVar, Dict, Type, Optional
+from warnings import warn
+
+from dyndis.util import similar, issubclass_tv
 
 Self = TypeVar('Self')
+
+type_aliases: Dict[type, Tuple[type, ...]] = {
+    float: (float, int),
+    complex: (float, int, complex)
+}
+
+
+def is_valid_args(args):
+    return not args or \
+           all(a in (Any, object) for a in args)
 
 
 def to_type_iter(t: Union[type, None], self_type):
@@ -15,23 +28,31 @@ def to_type_iter(t: Union[type, None], self_type):
     * types
     * singletons (..., None, Notimplemented)
     * the typing.Any object (equivalent to object)
+    * the dyndis.Self object
     * any non-specific typing abstract class (Sized, Iterable, ect...)
     * typing.Union
     """
     if isinstance(t, type):
-        return t,
+        alias = type_aliases.get(t, None)
+        return alias or (t,)
     if t is Any:
-        return to_type_iter(object, self_type)
+        return t,
     if t is Self:
         if self_type is _missing:
             raise ValueError('Self cannot be used as a type hint outside of Implementor')
         return to_type_iter(self_type, self_type)
     if t in (..., NotImplemented, None):
         return to_type_iter(type(t), self_type)
+    if isinstance(t, TypeVar):
+        if t.__contravariant__ or t.__covariant__:
+            raise ValueError(f'cannot use covariant or contravariant type hint {t}')
+        return t,
     if getattr(t, '__origin__', None) is Union:
         return tuple(chain.from_iterable(to_type_iter(a, self_type) for a in t.__args__))
-    if isinstance(getattr(t, '__origin__', None), type) and not t.__args__:
+    if isinstance(getattr(t, '__origin__', None), type) and is_valid_args(t.__args__):
         return to_type_iter(t.__origin__, self_type)
+    if getattr(t, '__origin__', None) is Callable.__origin__ and t.__args__[0] is ... and is_valid_args(t.__args__[1:]):
+        return to_type_iter(Callable, self_type)
     raise TypeError(f'type annotation {t} is not a type, give it a default to ignore it from the candidate list')
 
 
@@ -56,7 +77,8 @@ class Candidate:
         self.__doc__ = getattr(func, '__doc__', None)
 
     @classmethod
-    def from_func(cls, priority, func, fallback_type_hint=_missing, self_type=_missing):
+    def from_func(cls, priority, func, fallback_type_hint=_missing, self_type=_missing,
+                  priority_adjust: Optional[Callable] = ...):
         """
         create a list of candidates from function using the function's type hints. ignores all parameters with default
         values, as well as variadic parameters or keyword-only parameters
@@ -68,21 +90,45 @@ class Candidate:
 
         :return: a list of candidates generated from the function
         """
+        if priority_adjust is ...:
+            def priority_adjust(original, types):
+                t_var_count = sum(isinstance(t, TypeVar) for t in types)
+                if t_var_count:
+                    original -= t_var_count / (t_var_count + 1)
+                return original
+        elif priority_adjust is None:
+            def priority_adjust(original, types):
+                return original
+
         type_hints = get_type_hints(func)
         params = signature(func).parameters
         type_iters = []
+        super_type_iters = [type_iters]
         p: Parameter
         for p in params.values():
-            if p.kind not in (p.POSITIONAL_ONLY, p.POSITIONAL_OR_KEYWORD) \
-                    or p.default is not p.empty:
+            if p.kind not in (p.POSITIONAL_ONLY, p.POSITIONAL_OR_KEYWORD):
                 break
             t = type_hints.get(p.name, fallback_type_hint)
             if t is _missing:
+                if p.default is p.empty:
+                    break
                 raise KeyError(p.name)
             i = to_type_iter(t, self_type)
+            if p.default is not p.empty:
+                default_type = type(p.default)
+                if default_type is not object \
+                        and not any(issubclass_tv(default_type, x) for x in i):
+                    if p.default is None:
+                        i = (default_type, *i)
+                    else:
+                        warn(f'default value for parameter {p.name} is not included in its annotations', stacklevel=3)
+                super_type_iters.append(list(type_iters))
             type_iters.append(i)
-        type_lists = product(*type_iters)
-        return [cls(tuple(types), func, priority) for types in type_lists]
+
+        type_lists = list(chain.from_iterable(product(*ti) for ti in super_type_iters))
+        return [cls(tuple(types), func,
+                    priority_adjust(priority, tuple(types))
+                    ) for types in type_lists]
 
     def __str__(self):
         return (self.__name__ or 'unnamed candidate') + '<' + ', '.join(n.__name__ for n in self.types) + '>'
@@ -135,25 +181,51 @@ class Candidate:
         return ret
 
 
+def cmp_type_hint(r: Union[Type, TypeVar], l: Union[Type, TypeVar]):
+    """
+    can return 4 values:
+    0 if they are identical
+    -1 if r <= l
+    1 if l <= r
+    None if they cannot be compared
+    """
+    if isinstance(r, TypeVar):
+        if r.__bound__:
+            return cmp_type_hint(r.__bound__, l)
+        elif r.__constraints__:
+            return similar(cmp_type_hint(c, l) for c in r.__constraints__)
+        else:
+            return cmp_type_hint(object, l)
+    elif isinstance(l, TypeVar):
+        i_cth = cmp_type_hint(l, r)
+        return i_cth and -i_cth
+    else:  # both are types
+        if r is l:
+            return 0
+        elif issubclass(r, l):
+            return -1
+        elif issubclass(l, r):
+            return 1
+        return None
+
+
 def cmp_key(rhs: Tuple[type, ...], lhs: Tuple[type, ...]):
     """
     check whether two type tuples are ordered in any way
     :return: -1 if rhs is a sub-key of lhs, 1 if lhs is a sub-key of rhs, 0 if the two keys cannot be compared
-     (it is an error to sent identical keys)
+     (it is an error to send identical keys)
     """
     ret = 0
     for r, l in zip(rhs, lhs):
-        if r is l:
+        i = cmp_type_hint(r, l)
+        if i is None:
+            return 0
+        if i == 0:
             continue
-        elif issubclass(r, l):
-            if ret == 1:
-                return 0
-            ret = -1
-        elif issubclass(l, r):
-            if ret == -1:
-                return 0
-            ret = 1
-        else:
+
+        if ret == 0:
+            ret = i
+        elif ret != i:
             return 0
     return ret