+++ dyndis/implementor.py
@@ -48,7 +48,7 @@ class Implementor:
             if self.multidispatch.__name__ is not None \
                     and candidate.__name__ is not None\
                     and self.multidispatch.__name__.strip('_') != candidate.__name__.strip('_'):
-                warn(f'implementor {candidate} in class {owner.__name__} has dissimilar name from multidispatch {self.multidispatch}')
+                warn(f'implementor {candidate} in class {owner.__name__} has dissimilar name from multidispatch {self.multidispatch}', stacklevel=5)
             return candidate
 
         if self.locked: